+++ R/convert.R
@@ -182,7 +182,7 @@ convert_up <- function(from, to, call = sys.call(-1L)) {
 }
 
 is_down_cast <- function(x, class) {
-  inherits(x, setdiff(class_dispatch(class), "S7_object"))
+  class_dispatch_extends(obj_dispatch(x), class_dispatch(class))
 }
 
 convert_down <- function(from, to, ...) {